support ruby 1.8; define ConnectionPool::TimedQueue instead of claiming top-level ::TimedQueue
tested on MRI 1.8.7-p358, MRI 1.9.3-p0, and JRuby 1.6.7
