+++ test/helper.rb
@@ -8,7 +8,7 @@ puts RUBY_DESCRIPTION
 class MiniTest::Unit::TestCase
 
   def async_test(time=0.5)
-    q = TimedQueue.new
+    q = ConnectionPool::TimedQueue.new
     yield Proc.new { q << nil }
     q.timed_pop(time)
   end
